Defines all necessary calculator processes together with interface and connects the two
All of essential calculation processes are implemented and made intractable
with via web page interface via overall handler function together with
defineExpression and evaluateExpression functions.
First overall handler function in order to react to user input is tasked
with interpreting it and calling one of two appropriate functions due to
each user interaction being of one of two types and each function being
tasked with handling one of them and is tasked with handling all things
necessary to enable appropriate interaction between two functions.
defineExpression function makes sure that each user input that isnt
request for evaluation of expression is properly recorded as part of
expression and also exposed/displayed as part of user interface .
While evaluateExpression is tasked with parsing expression and passing it
to calculate function in order to fully evaluate it and return result.